Refactor team mongo
